requirements.txt
mypy
requests
black
isort
pyflakes
Django
djangorestframework
psycopg2
piprot
